Initial migration to GitHub
Edited sythesis of source commit messages follows:

---

Rely on the user setting `-DKatana_DIR` to the `plugin_apis/cmake`
directory in the Katana installation, and assume `katana-config.cmake`
will then configure any necessary CMake paths.

---

The plugin code is written against OpenAssetIO v1.0.0-rc.1.0 and
OpenAssetIO-MediaCreation v1.0.0-alpha.11.

---

Assume Katana already provides a Python environment for Python
OpenAssetIO plugins, which is true by default for both UI and
renderboot.

---

Use namespaced `KATANAOPENASSETIO_ENABLE_*` CMake cache variable
standard (similar to other OpenAssetIO projects) for enabling features.
I.e. warnings as errors, and hardening build flags.

---

Declare inline string constants, e.g. `std::string kAssetId`. We could
use a `constexpr std::string_view` instead, but that would require
constructing a `std::string` when it's used in a `std::string` context
(e.g. a dictionary key).

---

Add limited explicit error case checking and reporting.

E.g. we use `createEntityReference` and allow it to throw an exception
if the entity reference is not valid.

E.g. we assume the manager plugin has the desired capabilities and allow
OpenAssetIO to throw if not. If there are capabilities that are required
for the plugin to be valid, then this should be verified at plugin
loading time.

(Technically, the manager could lie and say they support a capability
without actually implementing the associated method, but this should be
considered a truly exceptional case and shouldn't be guarded against by
the host)

Both entry points for `resolveAsset` on the consumer side,
`FnDefaultAssetPlugin` and `FnAssetPlugin`, wrap exceptions such that
errors are re-thrown if `throwOnError` is `true` (very rare), and
returns the `assetId` unmodified if `throwOnError` is `false`.

In general, exceptions are usually just ignored in Katana, sadly.
Logging of errors may well be useful, but we should probably log all
errors, rather than particular exceptions. For now, allow exceptions to
propagate and let the host decide how they should be handled.

If more error information is needed then it would be better to finesse
the upstream OpenAssetIO exception messages than add explicit formatting
in the plugin.

---

Katana assumes that asset identifiers can be deconstructed into
individual fields, which can then be reconstituted into an identifier
(potentially, though rarely, after mutating a field).

This is what the AssetAPI `getAssetFields` function is referring to.
In particular, it is not meant to return a comprehensive dictionary of
all properties associated with the asset.

Instead, the complete list of properties is meant to be provided by
`getAssetAttributes` (filtered by some `scope` - usage of which in
the OpenAssetIO case needs future work to consider).

Therefore, the `resolve` of all `entityTraits` for an entity is via the
`getAssetAttributes` method. The `getAssetFields` method then only
`resolve`s/returns the required properties (`"name"` and `"version"`),
plus the entity reference, for use when the dictionary is passed into
`buildAssetId`.

---

The `FileUrlPathConverter` is not cheap to construct (on construction
it compiles a lot of regexes). Once constructed, it is stateless and
so safe to use multithreaded. So construct a single instance as a member
variable and re-use when needed.

---

Versioning matches Katana's expectations.

Based mostly on the Importomatic and LookFileManager nodes we can infer
that

* `resolveAssetVersion` should return the "stableTag" property of the
  Version trait of an entity.  I.e. "v2" rather than "latest".
* `getAssetVersions` should return a complete list of versions for an
  entity, including meta-versions, and the versions should be tags
  rather than full entity references.
* `buildAssetId` should use expect a version tag in the "version" input
  field, and use it to specify the version that the returned entity
  reference points to.

Support the `versionStr` argument of `resolveAssetVersion`, which,
according to the docstrings (doesn't seem to be used within Katana
anywhere), is a way to resolve a meta-version of an asset into a stable
version (e.g. "latest" to "v2").

There is overlap with this functionality and `buildAssetId`, so factor
out the common code.

---

Use the SourcePath trait to provide a scene graph location for a given
entity. It's not really clear if this is a good usage of the SourcePath
trait - the docstring seems to imply that it refers to a path _within_
the entity, rather than the entity's global path. However, it is the
best match for now, until we can clarify further.

Refs. LookFileOverrideEnable + LookFileResolve, and/or LookFileManager.

---

Make use of the hybrid plugin system, composing the Python plugin system
and the C++ plugin system.

---

Do the (re)initialization on `reset()`, and call `reset()` in the
constructor. The manager plugin will be loaded eagerly, rather than
lazily. In practice, this is fine - Katana uses the AssetAPI immediately
on opening a project, i.e. immediately after AssetAPI plugin
construction/`reset()`.

---

Add a `;` after `FnLogSetup`, or clang-format gets confused.

---

Append the (stable) version tag to the scene graph location returned
from `getUniqueScenegraphLocationFromAssetId` if the `includeVersion`
parameter is `true`.

Ref. LookFileOverrideEnable.

---

Use `(void)` trick rather than `[[maybe_unused]]` to silence various
unused variable warnings, where we expect that these variables will be
used in the near future.

---

Ensure we use a single `Context` for everything (consider
`createChildContext` if a new context is ever needed).

For publishing, use something akin to the Strategy pattern.

---

Add a super simple "Asset browser"

Add a simple UI delegate, whose sole functionality is to provide an
extra "Asset" tab in the asset browser. This tab simply presents a text
box where asset IDs can be typed/pasted, circumventing the File tab's
validation. Useful for testing. Disabled by default.

---

This is a library meant for runtime loading by a plugin system, so in
CMake parlance is a MODULE.

---

Rather than risk setting compiler flags globally, potentially affecting
consuming projects (i.e. via `add_subdirectory`), set them per target.

---

Disable Python plugin support if the `KATANAOPENASSETIO_DISABLE_PYTHON`
environment variable is set.

---

Support Katana 8+ in simple UI delegate. Katana 8+ uses PySide6 rather
than PyQt.

---

Apache2 license and copyright. Diverge from the OpenAssetIO standard to
use a more common/correct structure. I.e. copyright before SPDX, project
tag above copyright, "(c)" in copyright string, and addition of a NOTICE
file.